Extract link_checking outside of site/lib.rs
